pyproject.toml

[tool.poetry]
name = "music-central"
version = "0.1.0"
description = "Central spot for playing music online and gathering things together"
authors = ["Robert Fornof"]
license = "MIT"
readme = "README.md"
[tool.poetry.scripts]
start = "my_module:main"
test = "tests:"
[tool.poetry.dependencies]
python = "^3.12"
mido = "^1.3.2"
aiohttp = "^3.9.5"
aiohttp-retry = "^2.8.3"
aiosignal = "^1.3.1"
altgraph = "^0.17.4"
async-timeout = "^4.0.3"
attrs = "^23.2.0"
blinker = "^1.8.2"
certifi = "^2024.7.4"
charset-normalizer = "^3.3.2"
click = "^8.1.7"
colorama = "^0.4.6"
exceptiongroup = "^1.2.1"
frozenlist = "^1.4.1"
idna = "^3.7"
iniconfig = "^2.0.0"
itsdangerous = "^2.2.0"
jinja2 = "^3.1.4"
macholib = "^1.16.3"
markupsafe = "^2.1.5"
midiutil = "^1.2.1"
multidict = "^6.0.5"
pefile = "^2023.2.7"
pluggy = "^1.5.0"
pyjwt = "^2.8.0"
pytest = "^8.2.2"
python-rtmidi = "^1.5.8"
pywin32-ctypes = "^0.2.2"
pyyaml = "^6.0.1"
requests = "^2.32.3"
tomli = "^2.0.1"
tornado = "^6.4.1"
twilio = "^9.2.3"
urllib3 = "^2.2.2"
werkzeug = "^3.0.3"
yarl = "^1.9.4"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
